requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
requests==2.32.3
mysqlclient==2.2.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask-Cors==5.0.0